requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP client
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Supabase client for storage operations
supabase>=2.0.0

# Whisper transcription (OpenAI version - easier to install on Windows)
openai-whisper>=20231117
# Required for audio processing
librosa>=0.10.0

# Note: FFmpeg must be installed separately on the system
# For Ubuntu/Debian: sudo apt-get install ffmpeg
# For macOS: brew install ffmpeg
# For Windows: Download from https://ffmpeg.org/download.html

# LangGraph for workflow orchestration
langgraph>=0.2.59
langchain-openai>=0.2.14
langchain-core>=0.3.28
langgraph-checkpoint-postgres>=2.0.8

# PostgreSQL for checkpointing (required by langgraph-checkpoint-postgres)
psycopg[binary]>=3.2.3
psycopg-pool>=3.2.4

# Server-Sent Events support
sse-starlette>=2.1.3

# SVG → PNG conversion for logo overlay in rendered videos.
# Backends tried in order (most reliable first):
#
#   Backend 0 — pre-rendered logo.png in utils/assets/ (zero deps — recommended for Windows)
#
#   Backend 1 — cairosvg (best SVG quality, needs native Cairo):
#     Linux : apt install libcairo2 && pip install cairosvg
#     macOS : brew install cairo    && pip install cairosvg
#     Windows: requires GTK runtime (https://cairosvg.org/documentation/)
cairosvg>=2.7.0
#
#   Backend 2 — svglib + reportlab (needs reportlab _renderPM C extension):
svglib>=1.5.1
reportlab>=4.0.0
#
#   Backend 3 — Pillow (pure Python, generates styled text logo; works everywhere):
Pillow>=10.0.0
